rename shared to lib
