+++ src/common/header/Header.tsx
@@ -13,7 +13,7 @@ import {
 } from '@chakra-ui/react';
 import { Link } from 'react-router-dom';
 
-import { useSessionStorage } from '../../shared/hooks/storage/useSessionStorage';
+import { useSessionStorage } from '../../lib/hooks/storage/useSessionStorage';
 
 import { NavButtons } from './components/NavButtons';
 import { Search } from './components/SearchBar';
