+++ src/common/sidebar/components/Card.tsx
@@ -3,7 +3,7 @@ import { Box, Text, Flex, VStack, IconButton } from '@chakra-ui/react';
 import { PropsWithChildren, useCallback } from 'react';
 import { Link, useParams } from 'react-router-dom';
 
-import { useSavedCourses } from '../../../shared/hooks/useSavedCourses';
+import { useSavedCourses } from '../../../lib/hooks/useSavedCourses';
 
 export interface CardProps {
   /**
