+++ src/common/sidebar/containers/SidebarContainer.tsx
@@ -2,7 +2,7 @@ import { Center, Box, Flex, Heading, HStack, Spinner } from '@chakra-ui/react';
 import { useMemo, useState } from 'react';
 import { Route, Routes } from 'react-router';
 
-import { Course, Term, useGetCourses, useSubjects } from '../../../shared/fetchers';
+import { Course, Term, useGetCourses, useSubjects } from '../../../lib/fetchers';
 import { CoursesList } from '../components/CoursesList';
 import { CustomHits } from '../components/SearchResults';
 import { SubjectsList } from '../components/SubjectsList';
