+++ src/lib/styles/CalendarStyles.scss
