+++ src/lib/utils/courses.ts
