+++ src/pages/calendar/components/Seats.tsx
@@ -1,6 +1,6 @@
 import { Box, Flex, Heading, Progress } from '@chakra-ui/react';
 
-import { Seat } from '../../../shared/fetchers';
+import { Seat } from '../../../lib/fetchers';
 
 export interface SeatsProps {
   seat?: Seat;
