+++ src/pages/calendar/containers/Content.tsx
@@ -4,8 +4,8 @@ import { Helmet } from 'react-helmet';
 import { MdDelete, MdAdd } from 'react-icons/md';
 import { useSearchParams } from 'react-router-dom';
 
-import { Term, useGetCourse } from '../../../shared/fetchers';
-import { useSavedCourses } from '../../../shared/hooks/useSavedCourses';
+import { Term, useGetCourse } from '../../../lib/fetchers';
+import { useSavedCourses } from '../../../lib/hooks/useSavedCourses';
 import { CourseInfo } from '../components/Course';
 
 import { SectionsContainer } from './Section';
