+++ src/pages/calendar/containers/Section.tsx
@@ -1,7 +1,7 @@
 import { Box, Center, Divider, Heading, Spinner } from '@chakra-ui/react';
 
-import { ClassScheduleListing, Seat, Term, useSeats, useSections } from '../../../shared/fetchers';
-import { getReadableTerm } from '../../../shared/utils/terms';
+import { ClassScheduleListing, Seat, Term, useSeats, useSections } from '../../../lib/fetchers';
+import { getReadableTerm } from '../../../lib/utils/terms';
 import { SectionInfo } from '../components/Section';
 
 function Sections({ sections, seats }: { sections: ClassScheduleListing[]; seats?: Seat[] | null }): JSX.Element {
