+++ src/pages/registration/containers/CourseContainer.tsx
@@ -4,8 +4,8 @@ import { Collapse } from '@chakra-ui/transition';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { useParams } from 'react-router';
 
-import { useSeats, useSections, Term, Seat } from '../../../shared/fetchers';
-import { SavedCourse } from '../../../shared/hooks/useSavedCourses';
+import { useSeats, useSections, Term, Seat } from '../../../lib/fetchers';
+import { SavedCourse } from '../../../lib/hooks/useSavedCourses';
 import { RegistrationMinimized } from '../components/RegistrationMinimized';
 import { RegistrationSection } from '../components/RegistrationSection';
