+++ src/pages/scheduler/containers/SchedulerContainer.tsx
@@ -2,7 +2,7 @@ import { Box, Flex } from '@chakra-ui/react';
 import { useMemo } from 'react';
 import { useParams } from 'react-router';
 
-import { useSavedCourses } from '../../../shared/hooks/useSavedCourses';
+import { useSavedCourses } from '../../../lib/hooks/useSavedCourses';
 import { SchedulerCalendar } from '../components/SchedulerCalendar';
 import { useCalendarEvents } from '../hooks/useCalendarEvents';
 import { CalendarEvent } from '../shared/types';
