+++ src/pages/scheduler/shared/constants.ts
@@ -1,4 +1,4 @@
-import { SectionType } from '../../../shared/fetchers';
+import { SectionType } from '../../../lib/fetchers';
 
 export const SECTION_TYPES: {
   sectionName: SectionType;
